# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Riogas Gestión API",
    "description": "Catálogo de APIs de GOYA (app 3). Backend NestJS (prefijo /api) + route handlers del front Next.\nGenerado por `pnpm docs:api` con un parser estático del código; no editar a mano — las descripciones y ejemplos van en docs/api/anotaciones.yaml.",
    "version": "0.1.0"
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreatePuestoDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "minimum": 1
          },
          "nombre": {
            "type": "string",
            "minLength": 1,
            "maxLength": 40
          },
          "departamentoId": {
            "type": "integer"
          },
          "direccion": {
            "type": "string",
            "maxLength": 100
          },
          "localidadId": {
            "type": "integer"
          },
          "zonaId": {
            "type": "integer"
          },
          "mail": {
            "type": "string",
            "format": "email",
            "maxLength": 100
          },
          "telefono": {
            "type": "string",
            "maxLength": 20
          },
          "propio": {
            "type": "string",
            "enum": [
              "S",
              "N"
            ]
          },
          "autopedido": {
            "type": "string",
            "enum": [
              "S",
              "N"
            ]
          },
          "fleteCobra": {
            "type": "string",
            "enum": [
              "S",
              "N"
            ]
          },
          "fleteCantidad": {
            "type": "string",
            "maxLength": 2
          },
          "horarios": {
            "type": "string",
            "maxLength": 200
          },
          "lat": {
            "type": "number",
            "maximum": 90,
            "description": "Lat y lng son \"todo o nada\": una sola coordenada no ubica nada y dejaría\nel mini mapa apuntando al Golfo de Guinea."
          },
          "lng": {
            "type": "number",
            "maximum": 180
          },
          "estado": {
            "type": "string",
            "enum": [
              "A",
              "P"
            ]
          }
        },
        "required": [
          "departamentoId",
          "id",
          "nombre"
        ],
        "description": "Alta de puesto.\n\n`id` es obligatorio y lo elige quien crea: la tabla `puesto` NO tiene\nautoincremental (los ids vienen del AS400: 1, 2, 4, 5 … 100, 1000), así que\nun create sin id explícito rompe.\n\nLos flags del legado viajan como 'S'/'N' de un caracter, no como booleanos."
      },
...
      "CreateZonaDto": {
        "type": "object",
        "properties": {
          "puestoId": {
            "type": "integer"
          },
          "nombre": {
            "type": "string",
            "minLength": 1,
            "maxLength": 60
          },
          "descripcion": {
            "type": "string",
            "maxLength": 200
          },
          "color": {
            "type": "string",
            "pattern": "^#[0-9a-fA-F]{6}$"
          },
          "tipoZona": {
            "type": "string"
          },
          "servicios": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "estado": {
            "type": "string"
          },
          "poligono": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PuntoPoligonoDto"
            },
            "minItems": 3
          }
        },
        "required": [
          "color",
          "nombre",
          "poligono",
          "puestoId",
          "servicios",
          "tipoZona"
        ]
      },
...
      "PuntoPoligonoDto": {
        "type": "object",
        "properties": {
          "lat": {
            "type": "number"
          },
          "lng": {
            "type": "number"
          }
        },
        "required": [
          "lat",
          "lng"
        ]
      },
...
      "UpdatePuestoDto": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "minLength": 1,
            "maxLength": 40
          },
          "departamentoId": {
            "type": "integer"
          },
          "direccion": {
            "type": "string",
            "maxLength": 100
          },
          "localidadId": {
            "type": "integer"
          },
          "zonaId": {
            "type": "integer"
          },
          "mail": {
            "type": "string",
            "format": "email",
            "maxLength": 100
          },
          "telefono": {
            "type": "string",
            "maxLength": 20
          },
          "propio": {
            "type": "string",
            "enum": [
              "S",
              "N"
            ]
          },
          "autopedido": {
            "type": "string",
            "enum": [
              "S",
              "N"
            ]
          },
          "fleteCobra": {
            "type": "string",
            "enum": [
              "S",
              "N"
            ]
          },
          "fleteCantidad": {
            "type": "string",
            "maxLength": 2
          },
          "horarios": {
            "type": "string",
            "maxLength": 200
          },
          "lat": {
            "type": "number",
            "maximum": 90
          },
          "lng": {
            "type": "number",
            "maximum": 180
          },
          "estado": {
            "type": "string",
            "enum": [
              "A",
              "P"
            ]
          }
        },
        "description": "Edición: mismos campos, todos opcionales, sin poder cambiar el id."
      },
...
      "UpdateZonaDto": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "minLength": 1,
            "maxLength": 60
          },
          "descripcion": {
            "type": "string",
            "maxLength": 200
          },
          "color": {
            "type": "string",
            "pattern": "^#[0-9a-fA-F]{6}$"
          },
          "tipoZona": {
            "type": "string"
          },
          "servicios": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "estado": {
            "type": "string"
          },
          "poligono": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PuntoPoligonoDto"
            },
            "minItems": 3
          }
        },
        "description": "Todos los campos opcionales (PATCH). puestoId no se cambia: una zona no\nse muda de puesto (se elimina y se crea en el otro)."
      }
    }
  },
  "paths": {
...
    "/api/puestos": {
      "get": {
        "operationId": "PuestosController_findAll",
        "tags": [
          "puestos"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Búsqueda libre: nombre, dirección, departamento o id."
          },
          {
            "name": "estado",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Estado del puesto: A=Activo / P=Pasivo."
          },
          {
            "name": "departamentoId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "conZona",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "con",
                "sin"
              ]
            },
            "description": "con | sin — según tenga zonas operativas asignadas."
          },
          {
            "name": "conMoviles",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "con",
                "sin"
              ]
            },
            "description": "con | sin — según tenga móviles asociados."
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "x-goya-origen": "backend (NestJS)",
        "x-goya-archivo": "backend/src/puestos/puestos.controller.ts",
        "x-goya-auth": "JWT (Bearer)",
        "summary": "Listado paginado de puestos, con filtros.",
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "PuestosController_create",
        "tags": [
          "puestos"
        ],
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "x-goya-origen": "backend (NestJS)",
        "x-goya-archivo": "backend/src/puestos/puestos.controller.ts",
        "x-goya-auth": "JWT (Bearer)",
        "summary": "Alta de puesto.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePuestoDto"
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/puestos/{id}": {
      "get": {
        "operationId": "PuestosController_findOne",
        "tags": [
          "puestos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "x-goya-origen": "backend (NestJS)",
        "x-goya-archivo": "backend/src/puestos/puestos.controller.ts",
        "x-goya-auth": "JWT (Bearer)",
        "summary": "Ficha de un puesto.",
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "PuestosController_update",
        "tags": [
          "puestos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "x-goya-origen": "backend (NestJS)",
        "x-goya-archivo": "backend/src/puestos/puestos.controller.ts",
        "x-goya-auth": "JWT (Bearer)",
        "summary": "Modificación parcial de un puesto.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePuestoDto"
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/zonas": {
      "get": {
        "operationId": "ZonasController_findAll",
        "tags": [
          "zonas"
        ],
        "parameters": [
          {
            "name": "puestoId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "description": "Sin puestoId → devuelve las zonas de TODOS los puestos."
          },
          {
            "name": "tipoZona",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "servicio",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "incluirArchivadas",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "'false' → solo ACTIVE. Default: trae todas (el front filtra por capas)."
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "x-goya-origen": "backend (NestJS)",
        "x-goya-archivo": "backend/src/zonas/zonas.controller.ts",
        "x-goya-auth": "JWT (Bearer)",
        "summary": "Listado de zonas operativas, con filtros.",
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "ZonasController_create",
        "tags": [
          "zonas"
        ],
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "x-goya-origen": "backend (NestJS)",
        "x-goya-archivo": "backend/src/zonas/zonas.controller.ts",
        "x-goya-auth": "JWT (Bearer)",
        "summary": "Alta de zona operativa con su polígono.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateZonaDto"
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/zonas/{id}": {
      "patch": {
        "operationId": "ZonasController_update",
        "tags": [
          "zonas"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "x-goya-origen": "backend (NestJS)",
        "x-goya-archivo": "backend/src/zonas/zonas.controller.ts",
        "x-goya-auth": "JWT (Bearer)",
        "summary": "Modificación parcial de una zona (el puesto no se cambia).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateZonaDto"
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "ZonasController_remove",
        "tags": [
          "zonas"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "x-goya-origen": "backend (NestJS)",
        "x-goya-archivo": "backend/src/zonas/zonas.controller.ts",
        "x-goya-auth": "JWT (Bearer)",
        "summary": "Elimina una zona operativa.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/zonas/{id}/duplicar": {
      "post": {
        "operationId": "ZonasController_duplicar",
        "tags": [
          "zonas"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "x-goya-origen": "backend (NestJS)",
        "x-goya-archivo": "backend/src/zonas/zonas.controller.ts",
        "x-goya-auth": "JWT (Bearer)",
        "summary": "Duplica una zona con su polígono.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/zonas/puestos": {
      "get": {
        "operationId": "ZonasController_puestos",
        "tags": [
          "zonas"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "x-goya-origen": "backend (NestJS)",
        "x-goya-archivo": "backend/src/zonas/zonas.controller.ts",
        "x-goya-auth": "JWT (Bearer)",
        "summary": "Puestos con zonas, para el selector de la pantalla de zonificación.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/zonas/sync/pull": {
      "post": {
        "operationId": "ZonasController_syncPull",
        "tags": [
          "zonas"
        ],
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "x-goya-origen": "backend (NestJS)",
        "x-goya-archivo": "backend/src/zonas/zonas.controller.ts",
        "x-goya-auth": "JWT (Bearer)",
        "summary": "Importa/actualiza desde track (carga inicial y re-sync manual).",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/zonas/sync/push": {
      "post": {
        "operationId": "ZonasController_syncPush",
        "tags": [
          "zonas"
        ],
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "x-goya-origen": "backend (NestJS)",
        "x-goya-archivo": "backend/src/zonas/zonas.controller.ts",
        "x-goya-auth": "JWT (Bearer)",
        "summary": "Reintenta el espejo de las zonas PENDING/ERROR.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/zonas/sync/webhook": {
      "post": {
        "operationId": "ZonasController_syncWebhook",
        "tags": [
          "zonas"
        ],
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "x-goya-origen": "backend (NestJS)",
        "x-goya-archivo": "backend/src/zonas/zonas.controller.ts",
        "x-goya-auth": "x-api-key (SyncApiKeyGuard)",
        "summary": "Entrada para track: upsert de zonas con shape de track (x-api-key).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "x-goya-tipo": "{ zonas?: TrackZona[] } | TrackZona"
              }
            }
          }
        }
      }
    }
  },
...
}